backend/requirements.txt
django>=4.2
mysqlclient>=2.2.0
# If you prefer PyMySQL, comment the above and uncomment below
# pymysql>=1.1.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
python-dotenv>=1.0.1 